% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_activities_gpt_validation.R
\name{eco_activities_gpt_validation}
\alias{eco_activities_gpt_validation}
\title{ChatGPT match validation between Ecoinvent activities and EU Taxonomy activities' description}
\usage{
eco_activities_gpt_validation(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe
}
\description{
ChatGPT match validation between Ecoinvent activities and EU Taxonomy activities' description
}
\examples{
ecoinvent_activities_with_nace <- ecoinvent_activities_with_nace(
  ep_companies_ei_activities |> dplyr::slice(4:6),
  isic_nace_mapper
)

eu_tax_activities_all <- eu_tax_activities_all(
  eu_tax_climate_mitigation, eu_tax_climate_adaptation,
  eu_tax_water_protection, eu_tax_circular_economy_transition,
  eu_tax_pollution_prevention, eu_tax_biodiversity_protection
)
eu_tax_activities_all_with_keywords <- eu_tax_activities_all_with_keywords(
  eu_tax_activities_all,
  eu_tax_activities_keywords
)

eco_activities_after_keywords_search <- eco_activities_after_keywords_search(
ecoinvent_activities_eu_tax_activities(
  ecoinvent_activities_with_nace, eu_tax_activities_all_with_keywords
))

output <- eco_activities_gpt_validation(eco_activities_after_keywords_search)
output
}
