% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoinvent_activities_eu_tax_activities.R
\name{ecoinvent_activities_eu_tax_activities}
\alias{ecoinvent_activities_eu_tax_activities}
\title{Mapping Ecoinvent activities (with NACE) to EU Taxonomy activities}
\usage{
ecoinvent_activities_eu_tax_activities(
  ecoinvent_activities_with_nace,
  eu_tax_activities_all_with_keywords
)
}
\arguments{
\item{ecoinvent_activities_with_nace}{A dataframe}

\item{eu_tax_activities_all_with_keywords}{A dataframe}
}
\value{
A dataframe
}
\description{
Mapping Ecoinvent activities (with NACE) to EU Taxonomy activities
}
\examples{

ecoinvent_activities_with_nace <- ecoinvent_activities_with_nace(
  ep_companies_ei_activities |> head(10),
  isic_nace_mapper
)

eu_tax_activities_all <- eu_tax_activities_all(
  eu_tax_climate_mitigation, eu_tax_climate_adaptation,
  eu_tax_water_protection, eu_tax_circular_economy_transition,
  eu_tax_pollution_prevention, eu_tax_biodiversity_protection
)
eu_tax_activities_all_with_keywords <- eu_tax_activities_all_with_keywords(
  eu_tax_activities_all, eu_tax_activities_keywords
)

output <- ecoinvent_activities_eu_tax_activities(
  ecoinvent_activities_with_nace,
  eu_tax_activities_all_with_keywords
)
output
}
