% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoinvent_activities_with_nace.R
\name{ecoinvent_activities_with_nace}
\alias{ecoinvent_activities_with_nace}
\title{Map NACE codes to Ecoinvent activities}
\usage{
ecoinvent_activities_with_nace(ep_companies_ei_activities, isic_nace_mapper)
}
\arguments{
\item{ep_companies_ei_activities}{\link{ep_companies_ei_activities}}

\item{isic_nace_mapper}{\link{isic_nace_mapper}}
}
\value{
A dataframe
}
\description{
Map NACE codes to Ecoinvent activities
}
\examples{
output <- ecoinvent_activities_with_nace(ep_companies_ei_activities, isic_nace_mapper)
output
}
