% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ep_companies_ei_activities}
\alias{ep_companies_ei_activities}
\title{Europages companies mapped with Ecoinvent activities}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 123 rows and 7 columns.
}
\usage{
ep_companies_ei_activities
}
\description{
ep_companies_ei_activities
}
\keyword{datasets}
