% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_companies_ei_activities_at_product_level.R
\name{ep_companies_ei_activities_at_company_level}
\alias{ep_companies_ei_activities_at_company_level}
\title{Outputs company-level columns of \code{ep_companies_ei_activities}}
\usage{
ep_companies_ei_activities_at_company_level(data)
}
\arguments{
\item{data}{\link{ep_companies_ei_activities}}
}
\value{
A dataframe
}
\description{
Outputs company-level columns of \code{ep_companies_ei_activities}
}
\examples{
output <- ep_companies_ei_activities_at_company_level(ep_companies_ei_activities)
output
}
