% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu_tax_activities_all.R
\name{eu_tax_activities_all}
\alias{eu_tax_activities_all}
\title{Calculate all EU Taxonomy activities}
\usage{
eu_tax_activities_all(
  eu_tax_climate_mitigation,
  eu_tax_climate_adaptation,
  eu_tax_water_protection,
  eu_tax_circular_economy_transition,
  eu_tax_pollution_prevention,
  eu_tax_biodiversity_protection
)
}
\arguments{
\item{eu_tax_climate_mitigation}{\link{eu_tax_climate_mitigation}}

\item{eu_tax_climate_adaptation}{\link{eu_tax_climate_adaptation}}

\item{eu_tax_water_protection}{\link{eu_tax_water_protection}}

\item{eu_tax_circular_economy_transition}{\link{eu_tax_circular_economy_transition}}

\item{eu_tax_pollution_prevention}{\link{eu_tax_pollution_prevention}}

\item{eu_tax_biodiversity_protection}{\link{eu_tax_biodiversity_protection}}
}
\value{
A dataframe
}
\description{
Calculate all EU Taxonomy activities
}
\examples{
output <- eu_tax_activities_all(
  eu_tax_climate_mitigation, eu_tax_climate_adaptation,
  eu_tax_water_protection, eu_tax_circular_economy_transition,
  eu_tax_pollution_prevention, eu_tax_biodiversity_protection
)
output
}
