% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu_tax_activities_all_with_keywords.R
\name{eu_tax_activities_all_with_keywords}
\alias{eu_tax_activities_all_with_keywords}
\title{Assign domain-specific keywords to EU Taxonomy activities}
\usage{
eu_tax_activities_all_with_keywords(
  eu_tax_activities_all,
  eu_tax_activities_keywords
)
}
\arguments{
\item{eu_tax_activities_all}{A dataframe}

\item{eu_tax_activities_keywords}{\link{eu_tax_activities_keywords}}
}
\value{
A dataframe
}
\description{
Assign domain-specific keywords to EU Taxonomy activities
}
\examples{
eu_tax_activities_all <- eu_tax_activities_all(
  eu_tax_climate_mitigation, eu_tax_climate_adaptation,
  eu_tax_water_protection, eu_tax_circular_economy_transition,
  eu_tax_pollution_prevention, eu_tax_biodiversity_protection
)
output <- eu_tax_activities_all_with_keywords(eu_tax_activities_all, eu_tax_activities_keywords)
output
}
