% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eu_tax_activities_keywords}
\alias{eu_tax_activities_keywords}
\title{EU Taxonomy activities keywords}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.
}
\usage{
eu_tax_activities_keywords
}
\description{
eu_tax_activities_keywords
}
\keyword{datasets}
