% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eu_tax_circular_economy_transition}
\alias{eu_tax_circular_economy_transition}
\title{EU Taxonomy Objective: Transition to a Circular Economy}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 12 columns.
}
\usage{
eu_tax_circular_economy_transition
}
\description{
eu_tax_circular_economy_transition
}
\keyword{datasets}
