% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eu_tax_climate_adaptation}
\alias{eu_tax_climate_adaptation}
\title{EU Taxonomy Objective: Climate change adaptation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 106 rows and 12 columns.
}
\usage{
eu_tax_climate_adaptation
}
\description{
eu_tax_climate_adaptation
}
\keyword{datasets}
