% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eu_tax_climate_mitigation}
\alias{eu_tax_climate_mitigation}
\title{EU Taxonomy Objective: Climate change mitigation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 101 rows and 12 columns.
}
\usage{
eu_tax_climate_mitigation
}
\description{
eu_tax_climate_mitigation
}
\keyword{datasets}
