% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eu_tax_pollution_prevention}
\alias{eu_tax_pollution_prevention}
\title{EU Taxonomy Objective: Pollution Prevention}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 12 columns.
}
\usage{
eu_tax_pollution_prevention
}
\description{
eu_tax_pollution_prevention
}
\keyword{datasets}
