% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic_nace_mapper}
\alias{isic_nace_mapper}
\title{ISIC-NACE Mapper}
\format{
An object of class \code{data.frame} with 1167 rows and 2 columns.
}
\usage{
isic_nace_mapper
}
\description{
isic_nace_mapper
}
\keyword{datasets}
