% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_eligible.R
\name{taxonomy_eligible}
\alias{taxonomy_eligible}
\title{Outputs EU Taxonomy-eligible results}
\usage{
taxonomy_eligible(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe
}
\description{
Outputs EU Taxonomy-eligible results
}
\examples{
data <- tibble::tibble(
  ep_product = c("plastic", "iron"),
  keyword_is_present = c(TRUE, FALSE),
  gpt_validation = c("Yes", "No")
)
output <- taxonomy_eligible(data)
output
}
