% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilt.eu.taxonomy-package.R
\docType{package}
\name{tilt.eu.taxonomy-package}
\alias{tilt.eu.taxonomy}
\alias{tilt.eu.taxonomy-package}
\title{tilt.eu.taxonomy: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
